requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
python-dotenv==1.0.1
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
requests>=2.31.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
rich>=13.0.0
